3204: Use the preference data store for all three account settings
