+++ app/src/main/java/com/keylesspalace/tusky/settings/AccountPreferenceDataStore.kt
@@ -1,0 +1,32 @@
+package com.keylesspalace.tusky.settings
+
+import androidx.preference.PreferenceDataStore
+import com.keylesspalace.tusky.db.AccountEntity
+import com.keylesspalace.tusky.db.AccountManager
+
+class AccountPreferenceDataStore(private val accountManager: AccountManager, private val account: AccountEntity): PreferenceDataStore() {
+    override fun getBoolean(key: String?, defValue: Boolean): Boolean {
+        return when(key) {
+            PrefKeys.ALWAYS_SHOW_SENSITIVE_MEDIA -> account.alwaysShowSensitiveMedia
+            PrefKeys.ALWAYS_OPEN_SPOILER -> account.alwaysOpenSpoiler
+            PrefKeys.MEDIA_PREVIEW_ENABLED -> account.mediaPreviewEnabled
+            else -> defValue
+        }
+    }
+
+    override fun putBoolean(key: String?, value: Boolean) {
+        if (key == null) {
+            return;
+        }
+
+        when (key) {
+            PrefKeys.ALWAYS_SHOW_SENSITIVE_MEDIA -> account.alwaysShowSensitiveMedia = value
+            PrefKeys.ALWAYS_OPEN_SPOILER -> account.alwaysOpenSpoiler = value
+            PrefKeys.MEDIA_PREVIEW_ENABLED -> account.mediaPreviewEnabled = value
+        }
+
+        accountManager.saveAccount(account)
+
+        // TODO this should also dispatch the event; conflicts with setOnPreferenceChangeListener being executed before
+    }
+}
